Populate diary meals foods tables
Adds JS function such that on page load the foods for each meal that
are already attached to a diary entry are populated in the appropriate
tables. Two functions were added to diary.js. populateMealsTables iterates
through the outer object, each meal, and passes the mealName and the meal
object to the second functionaddFoodstoMealTable. Second function iterates
through the foods of the meal and attaches them to the appropriate meal
table. Page behavior is appropriate for function at this time.