Complete updating of total calories on page load
Modifies index.html to have total calories for each meal table
populated with 0. At present, population of the meal tables
on load of page will also adjust the total calories value with
each addition via the updateMealTotalCalories function. Same
takes argument of the meal name and the food object which is
being added to the table; it then takes the current value of
the total calories of the meal and increments it with the
calories of the food that is being added prior to updating
the text in the table cell. Page behavior is as anticipated
except for known bug in what is presently returning from
API call on page load. Feature complete.