Complete search foods by name feature
Adds jQuery listener to filter-foods input. On input change,
searchFoods function is triggered which clears the diary-foods
table and uses the makeFoodTable function to populate the table
with search results from ajax call to search endpoint. On full
deletion of the input, all foods are populated once again. Search
function is case insensitive as per user story. Page behavior is
remaining as desired.