Complete display/update of totals table
Modifies index.html to have appropriate pre-populated values
for Goal Calories, Calories Consumed, and Remaining Calories.
Adds updateConsumedCalories and updateRemainingCalories
functions which function as named.  Adds an invocation
of these functions on initial population of the meal
tables on page load as well as on addition of a food
to a meal table. All page behavior remains appropriate
at this time.