Complete remove food from meal function
Modifies addition of foods to meals such that a button is added with the meal_id
held in the data attribute of the button. On click of the button, a delete request
is sent to the appropriate API endpoint removing the food from the meal. Click
of the button also triggers functions which retrieve the meal name and calories
for the food from the DOM and passes them to functions to update calories total
table. Click of the button also triggers an update of the calories in the appropriate
meal table. Local confirmation indicates appropriate persistence of the deletion in
the database; values of the calorie totals also persist across refreshes. Page
behavior is appropriate at this time.