Adds requested refactors
Adds contructor to diary class which calls getDiary
and new diary listeners function on construction.
diaryListeners function is an instance method for
the diary class which adds event listeners for
navigation between diary dates. Also moves
diaryFoodTable.getFoods invocation to construction
of diary(this triggers the building of the foods
table on the diary page). All page functionality
remains as before.